✨ feat: SearchDialog 컴포넌트 필터링 및 강의 목록 표시 기능 개선
- SearchFilters 및 LectureTable 컴포넌트를 추가하여 검색 조건 및 강의 목록 표시를 분리
- SearchDialog에서 필터링 로직을 SearchFilters로 이동하여 코드 가독성 향상
- LectureTable을 통해 강의 목록을 효율적으로 렌더링하고 추가 기능 구현
- 불필요한 코드 제거 및 컴포넌트 구조 개선으로 유지보수성 향상
